+++ bump-version-local.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
-gsed -i.bak "/^(def version/c\
+gsed -i "/^(def version/c\
 (def version \"$1\")" src/hkimjp/wil2/view.clj